railway-deployment/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
llama-cpp-python>=0.2.20
python-multipart
